dog-expert-python/requirements.txt
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Pillow==7.0.0
tensorflow==2.7.0
